# docs/explore/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SozoRock Place Intelligence API",
    "version": "2026-08-08",
    "description": "Versioned, non-clinical county evidence, planning-agent and collaboration contracts. Missing evidence remains missing; county evidence is never represented as ZIP-, city- or person-level evidence."
  },
  "servers": [
    {
      "url": "https://health.sozorockfoundation.org"
    }
  ],
...
  "paths": {
...
    "/api/explore": {
      "get": {
        "tags": [
          "Evidence"
        ],
        "summary": "Resolve county, city/place or ZIP input and return the existing Explore presentation contract.",
        "responses": {
          "200": {
            "description": "Explore response."
          },
          "409": {
            "description": "Transparent county selection required."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cognitoAccessToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Cognito access token"
      }
    },
...
  },
...
}